moved unit tests into unit directory + updated imports
