+++ src/applications/ivc-champva/10-10D/tests/unit/components/CustomCheckboxRadioReviewPage.unit.spec.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { CustomCheckboxRadioReviewPage } from '../../../shared/components/CustomCheckboxRadioReviewPage';
-import { testComponentRender } from '../../../shared/tests/pages/pageTests.spec';
+import { CustomCheckboxRadioReviewPage } from '../../../../shared/components/CustomCheckboxRadioReviewPage';
+import { testComponentRender } from '../../../../shared/tests/pages/pageTests.spec';
 import mockData from '../fixtures/data/test-data.json';
-import { generateOptions } from '../../pages/ApplicantMedicareStatusPage';
+import { generateOptions } from '../../../pages/ApplicantMedicareStatusPage';
 
 testComponentRender(
   'CustomCheckboxRadioReviewPage',
