+++ src/applications/ivc-champva/10-10D/tests/unit/components/File/MissingFileOverview.unit.spec.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { expect } from 'chai';
-import SupportingDocumentsPage from '../../../pages/SupportingDocumentsPage';
-import { testComponentRender } from '../../../../shared/tests/pages/pageTests.spec';
+import SupportingDocumentsPage from '../../../../pages/SupportingDocumentsPage';
+import { testComponentRender } from '../../../../../shared/tests/pages/pageTests.spec';
 import {
   hasReq,
   checkFlags,
-} from '../../../../shared/components/fileUploads/MissingFileOverview';
+} from '../../../../../shared/components/fileUploads/MissingFileOverview';
 
-import formConfig from '../../../config/form';
+import formConfig from '../../../../config/form';
 import mockData from '../../fixtures/data/test-data.json';
 
 describe('hasReq', () => {
