+++ src/applications/ivc-champva/10-10D/tests/unit/config/signatureKeyName.unit.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import getNameKeyForSignature from '../../helpers/signatureKeyName';
+import getNameKeyForSignature from '../../../helpers/signatureKeyName';
 
 describe('Signature key name', () => {
   it('should be "veteransFullName" when certifierRole is "sponsor"', () => {
