+++ src/applications/ivc-champva/10-10D/tests/unit/helpers/helpers.unit.spec.js
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
 import React from 'react';
-import { applicantWording, getAgeInYears } from '../../../shared/utilities';
-import { isInRange } from '../../helpers/utilities';
-import ApplicantField from '../../../shared/components/applicantLists/ApplicantField';
-import { testComponentRender } from '../../../shared/tests/pages/pageTests.spec';
+import { applicantWording, getAgeInYears } from '../../../../shared/utilities';
+import { isInRange } from '../../../helpers/utilities';
+import ApplicantField from '../../../../shared/components/applicantLists/ApplicantField';
+import { testComponentRender } from '../../../../shared/tests/pages/pageTests.spec';
 import mockData from '../fixtures/data/test-data.json';
 
 describe('applicantWording helper', () => {
